+++ src/main/java/underground/atm/client/Client.java
@@ -1,0 +1,8 @@
+package underground.atm.client;
+
+import underground.atm.common.Request;
+import underground.atm.common.Response;
+
+public interface Client {
+    Response send(Request request);
+}
