+++ src/main/java/underground/atm/common/ResponseCodec.java
@@ -1,7 +1,5 @@
 package underground.atm.common;
 
-import underground.atm.server.CreditCardCodec;
-
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -13,11 +11,11 @@ public class ResponseCodec {
         switch (response) {
             case Response.AuthorizeResponse() -> dataOutput.write(Response.AuthorizeResponse.TYPE);
             case Response.DepositResponse() -> dataOutput.write(Response.DepositResponse.TYPE);
-            case Response.WithDrawResponse() -> dataOutput.write(Response.WithDrawResponse.TYPE);
+            case Response.WithdrawResponse() -> dataOutput.write(Response.WithdrawResponse.TYPE);
             case Response.TransferResponse() -> dataOutput.write(Response.TransferResponse.TYPE);
             case Response.ViewBalanceResponse(int balance) -> {
                 dataOutput.write(Response.ViewBalanceResponse.TYPE);
-                dataOutput.write(balance);
+                dataOutput.writeInt(balance);
             }
             case Response.ErrorResponse(int exceptionType) -> dataOutput.write(exceptionType);
         }
@@ -28,7 +26,7 @@ public class ResponseCodec {
         return switch (type) {
             case Response.AuthorizeResponse.TYPE -> { yield new Response.AuthorizeResponse();}
             case Response.DepositResponse.TYPE -> { yield new Response.DepositResponse();}
-            case Response.WithDrawResponse.TYPE -> { yield new Response.WithDrawResponse();}
+            case Response.WithdrawResponse.TYPE -> { yield new Response.WithdrawResponse();}
             case Response.TransferResponse.TYPE -> { yield  new Response.TransferResponse();}
             case Response.ViewBalanceResponse.TYPE -> {
                 int balance = inputStream.readInt();
