+++ src/main/java/underground/atm/common/data/CreditCard.java
@@ -1,0 +1,7 @@
+package underground.atm.common.data;
+
+public record CreditCard(int id, String pin, int amount) {
+    public CreditCard withAmount(int amount) {
+        return new CreditCard(id, pin, amount);
+    }
+}
