+++ src/main/java/underground/atm/common/exceptions/exceptions/AuthorizationFailedException.java
@@ -1,0 +1,5 @@
+package underground.atm.common.exceptions.exceptions;
+
+public final class AuthorizationFailedException extends Throwable {
+
+}
