+++ src/main/java/underground/atm/common/exceptions/exceptions/InvalidAmountException.java
@@ -1,0 +1,4 @@
+package underground.atm.common.exceptions.exceptions;
+
+public final class InvalidAmountException extends Throwable {
+}
