+++ src/main/java/underground/atm/common/exceptions/exceptions/NotEnoughMoneyException.java
@@ -1,0 +1,4 @@
+package underground.atm.common.exceptions.exceptions;
+
+public final class NotEnoughMoneyException extends Throwable {
+}
