--- src/main/java/underground/atm/data/CreditCard.java
@@ -1,7 +1,0 @@
-package underground.atm.data;
-
-public record CreditCard(int id, String pin, int amount) {
-    public CreditCard withAmount(int amount) {
-        return new CreditCard(id, pin, amount);
-    }
-}
