--- src/main/java/underground/atm/exceptions/AuthorizationFailedException.java
@@ -1,5 +1,0 @@
-package underground.atm.exceptions;
-
-public final class AuthorizationFailedException extends Throwable {
-
-}
