--- src/main/java/underground/atm/exceptions/CreditCardNotFoundException.java
@@ -1,4 +1,0 @@
-package underground.atm.exceptions;
-
-public final class CreditCardNotFoundException extends Throwable {
-}
