--- src/main/java/underground/atm/exceptions/InvalidAmountException.java
@@ -1,4 +1,0 @@
-package underground.atm.exceptions;
-
-public final class InvalidAmountException extends Throwable {
-}
