--- src/main/java/underground/atm/exceptions/NotEnoughMoneyException.java
@@ -1,4 +1,0 @@
-package underground.atm.exceptions;
-
-public final class NotEnoughMoneyException extends Throwable {
-}
