--- src/main/java/underground/atm/repositories/CreditCardDataSource.java
@@ -1,10 +1,0 @@
-package underground.atm.repositories;
-
-import underground.atm.data.CreditCard;
-
-import java.util.Collection;
-
-public interface CreditCardDataSource {
-    Collection<CreditCard> load();
-    void save(Collection<CreditCard> creditCreditCards);
-}
