--- src/main/java/underground/atm/repositories/Server_CreditCardRepository.java
@@ -1,11 +1,0 @@
-package underground.atm.repositories;
-
-import underground.atm.data.CreditCard;
-
-public interface Server_CreditCardRepository {
-    CreditCard findCardBy(int creditID);
-
-    void updateAmount(int cardID, int amount);
-
-
-}
