--- src/main/java/underground/atm/server/Client.java
@@ -1,5 +1,0 @@
-package underground.atm.server;
-
-public interface Client {
-    Response send(Request request);
-}
