+++ src/main/java/underground/atm/server/repositories/CreditCardDataSource.java
@@ -1,0 +1,10 @@
+package underground.atm.server.repositories;
+
+import underground.atm.common.data.CreditCard;
+
+import java.util.Collection;
+
+public interface CreditCardDataSource {
+    Collection<CreditCard> load();
+    void save(Collection<CreditCard> creditCreditCards);
+}
