+++ src/main/java/underground/atm/server/repositories/Server_CreditCardRepository.java
@@ -1,0 +1,11 @@
+package underground.atm.server.repositories;
+
+import underground.atm.common.data.CreditCard;
+
+public interface Server_CreditCardRepository {
+    CreditCard findCardBy(int creditID);
+
+    void updateAmount(int cardID, int amount);
+
+
+}
