--- src/test/java/underground/atm/server/ResponseCodecTest.java
@@ -1,28 +1,0 @@
-package underground.atm.server;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import underground.atm.data.CreditCard;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ResponseCodecTest {
-
-    private ResponseCodec responseCodec = new ResponseCodec();
-
-
-    @Test
-    @DisplayName("Encode Response Test")
-    void encodeResponseTest() {
-
-//        responseCodec.encode();
-
-//        responseCodec.decode .var
-    }
-
-}
